+++ web/js/033-utilities.js
@@ -1188,7 +1188,7 @@
 		}
 	};
 
-	Utilities.prototype.albumButtonWidth = function(thumbnailWidth) {
+	Utilities.albumButtonWidth = function(thumbnailWidth) {
 		if (env.options.albums_slide_style) {
 			return Math.ceil(thumbnailWidth * (1 + 2 * env.slideMarginFactor) + 2 * env.slideAlbumButtonBorder + 2 * env.slideBorder);
 		} else {
@@ -1196,7 +1196,7 @@
 		}
 	};
 
-	Utilities.prototype.thumbnailWidth = function(albumButtonWidth) {
+	Utilities.thumbnailWidth = function(albumButtonWidth) {
 		if (env.options.albums_slide_style) {
 			return Math.floor((albumButtonWidth - 2 * env.slideAlbumButtonBorder - 2 * env.slideBorder) / (1 + 2 * env.slideMarginFactor));
 		} else {
@@ -1385,7 +1385,8 @@
 			function(resolve_addAllSubalbums) {
 				var subalbumsPromises = [];
 				for (let indexSubalbum = self.subalbums.length - 1; indexSubalbum >= 0; indexSubalbum --) {
-					let addSubalbumPromise = self.addSubalbumToSelection(indexSubalbum, "#subalbum-select-box-" + indexSubalbum);
+					let id = PhotoFloat.convertCacheBaseToId(self.subalbums[indexSubalbum].cacheBase);
+					let addSubalbumPromise = self.addSubalbumToSelection(indexSubalbum, "#subalbum-select-box-" + id);
 					subalbumsPromises.push(addSubalbumPromise);
 				}
 				Promise.all(subalbumsPromises).then(
@@ -1405,7 +1406,8 @@
 					let subalbumsPromises = [];
 					for (let indexSubalbum = self.subalbums.length - 1; indexSubalbum >= 0; indexSubalbum --) {
 						// let subalbum = self.subalbums[indexSubalbum];
-						let removeSubalbumPromise = self.removeSubalbumFromSelection(indexSubalbum, "#subalbum-select-box-" + indexSubalbum);
+						let id = PhotoFloat.convertCacheBaseToId(self.subalbums[indexSubalbum].cacheBase);
+						let removeSubalbumPromise = self.removeSubalbumFromSelection("#subalbum-select-box-" + id);
 						subalbumsPromises.push(removeSubalbumPromise);
 					}
 					Promise.all(subalbumsPromises).then(
@@ -1540,7 +1542,9 @@
 					if (env.currentAlbum.isAlbumWithOneMedia()) {
 						// only one media has remained after the removal
 						env.currentMedia = env.currentAlbum.media[0];
-						env.currentMedia.show();
+						env.currentMediaIndex = 0;
+						$("#media-view-container").removeClass("hidden");
+						env.currentMedia.show(env.currentAlbum, "center");
 						// env.currentAlbum.prepareForShowing(0);
 					} else {
 						// more than one media has remained after the removal: remove the single media thumbnail
@@ -1559,7 +1563,9 @@
 						if (env.currentAlbum.isAlbumWithOneMedia()) {
 							// only one media has remained after the removal
 							env.currentMedia = env.currentAlbum.media[0];
-							env.currentAlbum.show();
+							env.currentMediaIndex = 0;
+							$("#media-view-container").removeClass("hidden");
+							env.currentMedia.show(env.currentAlbum, "center");
 							// env.currentAlbum.prepareForShowing(0);
 						} else if (clickedMediaIndex < env.currentMediaIndex) {
 							env.currentMediaIndex --;
@@ -1699,8 +1705,10 @@
 		);
 	};
 
-	Album.prototype.removeSubalbumFromSelection = function(iSubalbum, clickedSelector) {
+	Album.prototype.removeSubalbumFromSelection = function(clickedSelector) {
+		var iSubalbum = this.subalbums.findIndex(subalbum => subalbum.cacheBase === $(clickedSelector).parent().parent().attr("id"))
 		var subalbum = this.subalbums[iSubalbum];
+		// var subalbum = this.subalbums[iSubalbum];
 		var self = this;
 		return new Promise(
 			function(resolve_removeSubalbum) {
@@ -1790,6 +1798,140 @@
 		);
 	};
 
+	Utilities.prototype.setTitleOptions = function() {
+		if (env.options.hide_title)
+			$(".title").addClass("hidden-by-option");
+		else
+			$(".title").removeClass("hidden-by-option");
+
+		if (! env.options.show_album_media_count)
+			$(".title-count").addClass("hidden-by-option");
+		else
+			$(".title-count").removeClass("hidden-by-option");
+
+		$(".title").css("font-size", env.options.title_font_size);
+		$(".title-anchor").css("color", env.options.title_color);
+		$(".title-anchor").hover(function() {
+			//mouse over
+			$(this).css("color", env.options.title_color_hover);
+		}, function() {
+			//mouse out
+			$(this).css("color", env.options.title_color);
+		});
+	};
+
+	Utilities.prototype.setMediaOptions = function() {
+		$(".media-name").css("color", env.options.media_name_color);
+
+		if (env.currentMedia !== null && ! Utilities.isPopup() || ! env.options.show_media_names_below_thumbs)
+			$(".thumb-and-caption-container .media-name").addClass("hidden-by-option");
+		else
+			$(".thumb-and-caption-container .media-name").removeClass("hidden-by-option");
+
+		$(".thumb-and-caption-container").css("margin-right", env.options.spacing.toString() + "px");
+		$(".thumb-and-caption-container").css("margin-bottom", env.options.spacing.toString() + "px");
+
+		if (env.options.hide_descriptions)
+			$(".media-description").addClass("hidden-by-option");
+		else
+			$(".media-description").removeClass("hidden-by-option");
+
+		if (env.options.hide_tags)
+			$(".media-tags").addClass("hidden-by-option");
+		else
+			$(".media-tags").removeClass("hidden-by-option");
+
+		if (env.options.hide_bottom_thumbnails) {
+			$("#album-view.media-view-container").addClass("hidden-by-option");
+		} else {
+			$("#album-view.media-view-container").removeClass("hidden-by-option");
+		}
+
+	};
+
+	Utilities.prototype.setSubalbumsOptions = function() {
+		let scrollBarWidth = window.innerWidth - document.body.clientWidth || 15;
+
+		// resize down the album buttons if they are too wide
+		let albumViewWidth =
+			$("body").width() -
+			parseInt($("#album-view").css("padding-left")) -
+			parseInt($("#album-view").css("padding-right")) -
+			scrollBarWidth;
+		env.correctedAlbumThumbSize = env.options.album_thumb_size;
+		let correctedAlbumButtonSize = Utilities.albumButtonWidth(env.options.album_thumb_size);
+		if (albumViewWidth / (correctedAlbumButtonSize + env.options.spacing) < env.options.min_album_thumbnail) {
+			env.correctedAlbumThumbSize = Math.floor(Utilities.thumbnailWidth(albumViewWidth / env.options.min_album_thumbnail - env.options.spacing)) - 1;
+			correctedAlbumButtonSize = Utilities.albumButtonWidth(env.correctedAlbumThumbSize);
+		}
+		let captionFontSize = Math.round(Utilities.em2px("body", 1) * env.correctedAlbumThumbSize / env.options.album_thumb_size);
+		let captionHeight = parseInt(captionFontSize * 1.1) + 1;
+		let margin = 0;
+		if (env.options.albums_slide_style)
+			margin = Math.round(env.correctedAlbumThumbSize * env.slideMarginFactor);
+
+		let buttonAndCaptionHeight = correctedAlbumButtonSize + captionHeight;
+
+		let slideBorder = 0;
+		if (env.options.albums_slide_style)
+			slideBorder = env.slideBorder;
+
+		if (env.currentAlbum.isFolder() && ! env.options.show_album_names_below_thumbs)
+			$(".album-name").addClass("hidden-by-option");
+		else
+			$(".album-name").removeClass("hidden-by-option");
+
+		if (env.options.albums_slide_style) {
+			$(".album-name").css("color", env.options.album_slide_name_color);
+			$(".album-button-and-caption").css("background-color", env.options.album_slide_background_color);
+			$(".album-button").css("background-color", env.options.album_slide_background_color);
+			$(".album-caption, .album-caption .real-name").css("color", env.options.album_slide_caption_color);
+			$(".album-button-and-caption").addClass("slide");
+		} else {
+			$(".album-name").css("color", env.options.album_name_color);
+			$(".album-button-and-caption").css("background-color", "");
+			$(".album-button").css("background-color", "");
+			$(".album-caption, .album-caption .real-name").css("color", env.options.album_caption_color);
+			$(".album-button").css("border", "none");
+			$(".album-button-and-caption").removeClass("slide");
+		}
+
+		let marginBottom = env.options.spacing;
+		if (! env.options.albums_slide_style)
+			marginBottom += Utilities.em2px("body", 2);
+		$(".album-button-and-caption").css("width", (correctedAlbumButtonSize - 2 * slideBorder) + "px");
+		$(".album-button-and-caption").css("height", buttonAndCaptionHeight + "px");
+		$(".album-button-and-caption").css("margin-right", env.options.spacing.toString() + "px");
+		$(".album-button-and-caption").css("margin-bottom", marginBottom.toString() + "px");
+
+		$(".album-button").css("margin", margin + "px");
+		$(".album-button").css("width", env.correctedAlbumThumbSize + "px");
+		$(".album-button").css("height", env.correctedAlbumThumbSize + "px");
+
+		$(".album-caption").css("width", env.correctedAlbumThumbSize + "px");
+		$(".album-caption").css("height", captionHeight + "px");
+		$(".album-caption").css("font-size", captionFontSize + "px");
+
+		$(".album-tags").css("font-size", (Math.round(captionFontSize * 0.75)) + "px");
+
+
+		if (! env.options.show_album_media_count)
+			$(".album-caption-count").addClass("hidden-by-option");
+		else
+			$(".album-caption-count").removeClass("hidden-by-option");
+
+
+		if (env.options.hide_descriptions)
+			$(".album-description").addClass("hidden-by-option");
+		else
+			$(".album-description").removeClass("hidden-by-option");
+
+		if (env.options.hide_tags)
+			$(".album-tags").addClass("hidden-by-option");
+		else
+			$(".album-tags").removeClass("hidden-by-option");
+	};
+
 	Utilities.addTagLink = function(tag) {
 		// tags can be phrases (e.g. with automatic tags from person recognition)
 
@@ -1802,7 +1944,7 @@
 		return "<a href='" + hash + "'>" + tag + "</a>";
 	};
 
-	Utilities.prototype.em2px = function(selector, em) {
+	Utilities.em2px = function(selector, em) {
 		var emSize = parseFloat($(selector).css("font-size"));
 		return (em * emSize);
 	};
@@ -2259,7 +2401,7 @@
 					}
 					Utilities.setSelectButtonPosition();
 					Utilities.correctPrevNextPosition();
-					Utilities.setDescriptionPosition();
+					Utilities.setDescriptionOptions();
 
 				}
 
@@ -3211,11 +3353,11 @@
 			if (property === "tags")
 				return (object[property] !== undefined && object[property].length);
 			else
-				return (object[property] !== undefined && object[property]);
+				return (object[property] !== undefined && object[property] !== "");
 		}
 
-		let hasTitle = (object.title !== undefined && object.title);
-		let hasDescription = (object.description !== undefined && object.description);
+		let hasTitle = (object.title !== undefined && object.title !== "");
+		let hasDescription = (object.description !== undefined && object.description !== "");
 		let hasTags = (object.tags !== undefined && object.tags.length);
 
 		return hasTitle || hasDescription || hasTags;
@@ -3280,7 +3422,22 @@
 		Utilities.setDescription(this.metadata);
 	};
 
-	Utilities.setDescriptionPosition = function() {
+	Utilities.setDescriptionOptions = function() {
+		if (env.options.hide_descriptions && env.options.hide_tags)
+			$("#description-wrapper").addClass("hidden-by-option");
+		else
+			$("#description-wrapper").removeClass("hidden-by-option");
+
+		if (env.options.hide_descriptions)
+			$("#description").addClass("hidden-by-option");
+		else
+			$("#description").removeClass("hidden-by-option");
+
+		if (env.options.hide_tags)
+			$("#description-tags").addClass("hidden-by-option");
+		else
+			$("#description-tags").removeClass("hidden-by-option");
+
 		var thumbsHeight = 0;
 		if (env.currentMedia !== null && $("#album-view").is(":visible"))
 			thumbsHeight = env.options.media_thumb_size + 20;
@@ -3369,7 +3526,7 @@
 				}
 				$("#description-hide, #description-show").toggle();
 				$("#description, #description-tags").toggle();
-				Utilities.setDescriptionPosition();
+				Utilities.setDescriptionOptions();
 			}
 		);
 	};
@@ -3763,7 +3920,7 @@
 	Utilities.prototype.setPinchButtonsPosition = Utilities.setPinchButtonsPosition;
 	Utilities.prototype.setSelectButtonPosition = Utilities.setSelectButtonPosition;
 	Utilities.prototype.correctPrevNextPosition = Utilities.correctPrevNextPosition;
-	Utilities.prototype.setDescriptionPosition = Utilities.setDescriptionPosition;
+	Utilities.prototype.setDescriptionOptions = Utilities.setDescriptionOptions;
 	Utilities.prototype.toggleMenu = Utilities.toggleMenu;
 
 	window.Utilities = Utilities;